done ==> BE: refactor feature | FE: init DiaryPage, DiaryDetailPage

+++ BE/feature/diary/src/main/java/com/example/diary/service/DiaryServiceImpl.java
@@ -1,13 +1,8 @@
 package com.example.diary.service;
 
-import com.example.common.dto.diary.request.CreateDiaryItemsReq;
-import com.example.common.dto.diary.request.CreateDiaryReq;
-import com.example.common.dto.diary.request.DiaryItemRequest;
-import com.example.common.dto.diary.request.UpdateDiaryRequest;
-import com.example.common.dto.diary.response.CreateDiaryItemsRes;
-import com.example.common.dto.diary.response.CreateDiaryRes;
-import com.example.common.dto.diary.response.GetDiaryDetailRes;
-import com.example.common.dto.diary.response.GetDiaryRes;
+import ch.qos.logback.core.spi.ErrorCodes;
+import com.example.common.dto.diary.request.*;
+import com.example.common.dto.diary.response.*;
 import com.example.common.enums.ErrorCode;
 import com.example.common.exception.AppException;
 import com.example.common.interfaces.diary.DiaryService;
@@ -16,10 +11,7 @@ import com.example.diary.config.DiarySpecification;
 import com.example.diary.mapper.DiaryMapper;
 import com.example.diary.repository.UserDiaryItemRepository;
 import com.example.diary.repository.UserDiaryRepository;
-import com.example.persistence.entity.PurchaseOrder;
-import com.example.persistence.entity.PurchaseOrderItem;
-import com.example.persistence.entity.UserDiary;
-import com.example.persistence.entity.UserDiaryItem;
+import com.example.persistence.entity.*;
 import com.example.persistence.enumTable.DiaryStatus;
 import com.example.security.Util.UtilSecurityClass;
 import jakarta.transaction.Transactional;
@@ -30,6 +22,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
@@ -73,13 +66,14 @@ public class DiaryServiceImpl implements DiaryService {
 
     @Override
     public Page<GetDiaryRes> getDiaries(
+            String userId,
             String keyword,
             String fromDate,
             String toDate,
             Pageable pageable
     ) {
         Specification<UserDiary> spec = DiarySpecification.from(
-                DiarySpecification.DiaryFilter.of(keyword, fromDate, toDate)
+                DiarySpecification.DiaryFilter.of(userId,keyword, fromDate, toDate)
         );
         return diaryRepository
                 .findAll(spec, pageable)
@@ -87,17 +81,20 @@ public class DiaryServiceImpl implements DiaryService {
     }
 
     @Override
-    public GetDiaryDetailRes getDiaryDetail (String id) {
-        UserDiary diary = findDiaryOrThrow(id);
+    public GetDiaryDetailRes getDiaryDetail (String diaryId) {
+        UserDiary diary = findDiaryOrThrow(diaryId);
         return mapper.toGetDiaryDetail(diary);
     }
 
 
     @Transactional
     @Override
-    public CreateDiaryRes createDiary(CreateDiaryReq request) {
+    public CreateDiaryRes createDiary(String userId, CreateDiaryReq request) {
+
+        User user = userInternalService.getUserById(userId);
 
         UserDiary diary = UserDiary.builder()
+                .user(user)
                 .diaryName(request.getDiaryName())
                 .diaryStatus(DiaryStatus.UNPAID)
                 .note(request.getNote())
@@ -139,40 +136,90 @@ public class DiaryServiceImpl implements DiaryService {
 
     @Override
     @Transactional
-    public CreateDiaryRes updateDiary(String id, UpdateDiaryRequest request) {
-//        UserDiary diary = findDiaryOrThrow(id);
-//
-//        mapper.updateDiaryFromRequest(request, diary);
-//
-//        if (request.items() != null && !request.items().isEmpty()) {
-//            itemRepository.deleteByDiaryId(diary.getId());
-//            List<UserDiaryItem> items = buildItems(diary, request.items());
-//            itemRepository.saveAll(items);
-//            diary.setTotalAmount(calcTotal(items));
-//            diary = diaryRepository.save(diary);
-//            log.info("=== DIARY UPDATED: id={}", diary.getId());
-//            return buildDiaryResponse(diary, items);
-//        }
-//
-//        diary = diaryRepository.save(diary);
-//        List<UserDiaryItem> items = itemRepository.findByDiaryId(diary.getId());
-//        log.info("=== DIARY UPDATED: id={}", diary.getId());
-//        return buildDiaryResponse(diary, items);
+    public UpdateDiaryRes updateDiary(String userId,
+                                      String diaryId,
+                                      UpdateDiaryReq request) {
 
-        return null;
+        userInternalService.getUserById(userId);
+
+        UserDiary userDiary = findDiaryOrThrow(diaryId);
+
+        if (!userDiary.getUser().getId().equals(userId)) {
+            throw new AppException(ErrorCode.DIARY_NOT_BELONG_TO_USER);
+        }
+
+        mapper.updateDiaryFromRequest(request, userDiary);
+
+        UserDiary updated = diaryRepository.save(userDiary);
+
+        return mapper.toUpdateDiaryRes(updated);
+    }
+
+    @Override
+    public UpdateItemRes updateItems(String diaryId,
+                                     String itemId,
+                                     UpdateItemsReq request) {
+
+        UserDiary userDiary = findDiaryOrThrow(diaryId);
+
+        UserDiaryItem item = itemRepository.findById(itemId)
+                .orElseThrow(() -> new AppException(ErrorCode.DIARY_ITEMS_NOT_FOUND));
+
+        if (!item.getDiary().getId().equals(diaryId)) {
+            throw new AccessDeniedException(ErrorCode.DIARY_ITEMS_NOT_FOUND.getMessage());
+        }
+
+        mapper.updateItemFromRequest(request, item);
+
+        itemRepository.save(item);
+
+        List<UserDiaryItem> allItems = itemRepository.findByDiaryId(diaryId);
+
+        reCalculateUserDiary(userDiary, allItems);
+
+        diaryRepository.save(userDiary);
+
+        return mapper.toUpdateItemRes(item);
+    }
+
+    @Override
+    public UpdateDiaryStatusRes updateDiaryStatus(String userId, String diaryId, UpdateDiaryStatusReq request) {
+
+        userInternalService.getUserById(userId);
+
+        UserDiary userDiary = findDiaryOrThrow(diaryId);
+
+        if (!userDiary.getUser().getId().equals(userId)) {
+            throw new AccessDeniedException(ErrorCode.DIARY_ITEMS_NOT_FOUND.getMessage());
+        }
+
+        userDiary.setDiaryStatus(request.getDiaryStatus());
+
+        diaryRepository.save(userDiary);
+
+        return mapper.toUpdateDiaryStatusRes(userDiary);
     }
 
     @Override
     @Transactional
-    public void deleteDiary(String id) {
-        findDiaryOrThrow(id);
-        itemRepository.deleteByDiaryId(id);
+    public void deleteDiary(String userId, String id) {
+        UserDiary userDiary = findDiaryOrThrow(id);
+
+        if (!userDiary.getUser().getId().equals(userId)) {
+            throw new AppException(ErrorCode.DIARY_NOT_BELONG_TO_USER);
+        }
+
         diaryRepository.deleteById(id);
     }
 
     @Override
     public void deleteDiaryItem(String diaryId, String itemId) {
 
+        itemRepository
+                .findById(itemId)
+                .orElseThrow(() -> new AppException(ErrorCode.DIARY_ITEMS_NOT_FOUND));
+
+        itemRepository.deleteById(itemId);
     }
 
 }
